requirements.txt
pydantic>=2.8.0
openai>=1.58.0
python-dotenv>=1.0.1
requests>=2.32.0 
singlestoredb
google-search-results
fastapi
uvicorn
websockets
langgraph
langchain
langchain_openai